refactor(step3): 메서드 분리
